Setup navigation.
