+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.NewsPaper"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".NewsActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
