+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/breakingNewsFragment"
+        android:title="@string/breaking_news"
+        android:icon="@drawable/ic_home"/>
+
+    <item
+        android:id="@+id/savedNewsFragment"
+        android:title="@string/saved_news"
+        android:icon="@drawable/ic_save"/>
+
+    <item
+        android:id="@+id/searchNewsFragment"
+        android:title="@string/search_news"
+        android:icon="@drawable/ic_search"/>
+
+</menu>
